+++ checks/tasks/__init__.py
@@ -1,27 +1,15 @@
 # Copyright: 2022, ECP, NLnet Labs and the Internet.nl contributors
 # SPDX-License-Identifier: Apache-2.0
 import os
-from typing import Dict, Optional
-
-import requests
 import socket
 import time
 
-import urllib3
+import unbound
 from celery import Task
 from celery.exceptions import SoftTimeLimitExceeded
 from django.conf import settings
 
-import unbound
-from forcediphttpsadapter.adapters import ForcedIPHTTPSAdapter
-
-from checks.tasks.tls_connection_exceptions import NoIpError
 from internetnl import log
-from timeit import default_timer as timer
-
-# Disable HTTPS warnings.
-# TODO: is this the best place for this call?
-urllib3.disable_warnings()
 
 
 class SetupUnboundContext(Task):
@@ -124,62 +112,3 @@ class SetupUnboundContext(Task):
                 data["data"] = result.data
                 data["rcode"] = result.rcode
         data["done"] = True
-
-    def http_get(
-        self, url: str, headers: Optional[Dict] = None, session: Optional[requests.Session] = None, *args, **kwargs
-    ) -> requests.Response:
-        """
-        Perform a HTTP GET request using the stored session
-        """
-        # TODO: auto-retry
-        # TODO: make this this runnable from outside a task
-        start_time = timer()
-
-        if not headers:
-            headers = {}
-        headers["User-Agent"] = "internetnl/1.0"
-        if not session:
-            session = requests.session()
-
-        response = session.get(url, headers=headers, stream=True, *args, **kwargs)
-        log.debug(f"HTTP request completed in {timer()-start_time:.06f}s: {url} (headers: {headers})")
-        return response
-
-    def http_get_ip(
-        self,
-        hostname: str,
-        ip: str,
-        port: int,
-        path: str = "/",
-        https: bool = True,
-        headers: Optional[Dict] = None,
-        *args,
-        **kwargs,
-    ) -> requests.Response:
-        path = path.lstrip("/")
-        if not headers:
-            headers = {}
-
-        session = requests.session()
-        if https:
-            session.mount(f"https://{hostname}", ForcedIPHTTPSAdapter(dest_ip=ip))
-            url = f"https://{hostname}:{port}/{path}"
-        else:
-            if ":" in ip:
-                ip = f"[{ip}]"
-            url = f"http://{ip}:{port}/{path}"
-        headers["Host"] = hostname
-        return self.http_get(url, verify=False, headers=headers, session=session, *args, **kwargs)
-
-    def http_get_af(self, hostname: str, port: int, af: socket.AddressFamily, *args, **kwargs) -> requests.Response:
-        rr_type = unbound.RR_TYPE_AAAA if af == socket.AF_INET6 else unbound.RR_TYPE_A
-        # cb_data = ub_resolve_with_timeout(host, rr_type, unbound.RR_CLASS_IN, timeout)
-        # ips = [socket.inet_ntop(af, rr) for rr in cb_data["data"].data]
-        ips = self.resolve(hostname, rr_type)
-        exc = NoIpError(f"Unable to resolve {rr_type} record for host '{hostname}'")
-        for ip in ips:
-            try:
-                return self.http_get_ip(hostname, ip, port, *args, **kwargs)
-            except requests.RequestException as request_exception:
-                exc = request_exception
-        raise exc
